// src/app/core/services/model/localized.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EDteFunctionTestLanguageKeys } from './eDteFunctionTestLanguageKeys';

export interface Localized { 
    i18N?: EDteFunctionTestLanguageKeys;
    /**
     * Default text
     */
    default?: string;
    /**
     * String formats
     */
    formats?: Array<string>;
}